backend: Update devices instantly
With the new work in place to remove cursor focus in Clutter when the
wayland-pointer is hidden, this is causing a problem when showing the
pointer again: The enter event being sent to the actor is deferred??

Another benefit of this is that it prevents the flickering of the OSK
that happens when moving the mouse after opening the keyboard with a
touchscreen.

This commit reverts commit 741dd674.