events: Don't bypass clutter by default
This is a big step towards a more Wayland-centric event delivery system,
it allows Clutter to handle events that interact with windows in Wayland
sessions. This, together with handling actor events before sending them
to Wayland surfaces should allow moving quite a bit of stuff from
meta_display_handle_event to event handlers of Clutter actors in the
future.

Right now we need this for the screen keyboard to get notified about
key-presses of a phsyical keyboard.