backend: Change last-device-changed signal to last-x11-device-changed
Rename the last-device-changed signal to last-x11-device-changed and
only emit the signal if we get notified about a change by the x11
backend. This allows the OSK to get notified if a phsical keyboard is
used in x11 sessions. The signal will get emitted a lot less often than
last-device-changed, which fixes the issue that the now reverted
741dd674 fixed before.