clutter/seat: Add method to peek list of input devices
Add a method to ClutterSeat that allows peeking the list of input
devices and allow looping through devices a bit faster.

To make sure we don't break compatibilty with the old list_devices()
API, implement that function using a simple g_list_copy() of the list
returned by peek_devices().