clutter/stage-cogl: Fix painting the redraw clip with the damage region
The redraw clip that's painted together with the damage region has to be
copied earlier than we do right now. That is because if
PAINT_DAMAGE_REGION is enabled, buffer age is disabled and thus
may_use_clipped_redraw is FALSE, which means the queued_redraw_clip is
going to be set to the full view-rect. Fix that by copying the redraw
clip a bit earlier.

Also keep the queued_redraw_clip around a bit longer so it can actually
be used by paint_damage_region() and isn't freed before that.

While at it, move paint_damage_region() from swap_framebuffer() into
clutter_stage_cogl_redraw_view().