clutter/stage-cogl: Check for DISABLE_CLIPPED_REDRAWS earlier
Right now we're checking for the DISABLE_CLIPPED_REDRAWS debug flag
after creating the fb_clip_region and adjusting the redraw_clip means
that if may_use_clipped_redraw was TRUE, the redraw_clip will still be
set to the actual changed region and thus cause the stage to be
partially redrawn. Since we don't push a clip to the framebuffer though,
parts of the view will get corrupted.

To fix that, disable clipped redraws right away if the debug flag is
set. This also allows removing the may_use_clipped_redraw bool and
replacing it entirely with use_clipped_redraw.