clutter/stage-cogl: Check for buffer age early
Check for the buffer age before setting up the fb_clip_region and the
redraw_clip and set use_clipped_redraw to FALSE if the buffer age is
invalid. This ensures the redraw_clip is set to correctly set to the
view rect because we want to force a full redraw.