clutter/actor: Simplify resource scale calculation a bit
A simple for-loop is enough here, no need for a separate helper
function, also we don't need to get the stage on every recursive pass.

We still have to check whether the actor is mapped when going up the
hierarchy since it's possible an actor is mapped while its parents are
not because ClutterClone temporarily forces an actor to be mapped when
painting it.