clutter/actor: Cache the complete transformation matrices of actors
If we want to cache the absolute transformation matrix of an actor
(the matrix relative to eye coordinates), we need to ensure the
actor-to-parent transformation matrices get properly invalidated. This
is currently only the case if the actor subclass does not override the
apply_transform() vfunc.

So provide a new public API to invalidate the transformation matrix for
actors implementing custom transformations. This allows us to cache the
whole matrix and not only the part that's put together using
ClutterActors internal matrices.

This is also going to be needed to invalidate the stage-views list
reliably on changes to the actors transformation matrix.