clutter/actor: Cache the complete transformation matrices of actors
If we want to cache the absolute transformation matrix of an actor
(the matrix relative to eye coordinates), we need to ensure the
actor-to-parent transformation matrices get properly invalidated. This
is currently only the case if the actor subclass does not override the
apply_transform() vfunc.

So provide a new API that allows invalidating the transformation matrix
for actors implementing custom transformations, too. This in turn allows
us to cache the whole matrix and not only the part that's put together
by Clutter.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1299