clutter/actor: Cache absolute modelview and projection matrix
Add caching of the complete modelview and projection matrix applied to
an actor (the matrix that transforms to stage window coordinates).

We need this matrix very often because it's used every time we want the
absolute coordinates or vertices of an actor, and that happens almost
twice on every paint cycle (sometimes for calculating the resource scale
and always for transforming the paint volume).

So start caching this matrix for every actor and introduce the necessary
invalidation logic for that.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1299