clutter/actor: Cache absolute modelview and projection matrix
Add caching of the complete modelview and projection matrix applied to
an actor (the matrix that transforms to opengl coordinates).

We need this matrix very often because it's used every time we want the
absolute coordinates or vertices of an actor, and that happens at least
once every paint cycle (calculating the stage-views an actor is on and
transforming the paint volume).

So start caching this matrix for every actor and introduce the necessary
invalidation logic for that.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1299