clutter/stage: Peek viewport instead of copying it
Copying the same viewport around all the time for every step when
calculating the transformation matrices of actors is quite a lot of
unnecessary work. We can avoid this work by peeking instead of copying
the viewport of the stage, and only copying it later if necessary.

So rename _clutter_stage_get_viewport() to clutter_stage_peek_viewport()
and simply return a pointer to the float array there.

Also, since this is private API only used inside Clutter, we can avoid
the type check here and assume the passed ClutterStage is valid.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1299