clutter/actor: Allow computing resource scale while unmapped
It should also be possible to come up with a resource scale for actors
when the actor is not mapped yet. This is important for actors like
ClutterText, which request the resource scale inside their
get_preferred_width/height() implementations and calling those is valid
while an actor isn't mapped, too.

So try a bit harder to come up with a valid resource scale in this case
and allow calculating the scale if the actor is unmapped, too. This does
not yet fix the scale used in ClutterTexts get_preferred_width/height()
implementations since we also disallow computing the resource scale
while inside a size negotiation function, we'll fix that in a later
commit.

This fixes a bug found by Jonas Ådahl where a ClutterClone tries to
paint an unmapped (but still realized) actor with a
ClutterOffscreenEffect. The offscreen effect requests the actors
resource scale and the assertion of the resource scale being valid fails
(the actor is unmapped), causing a crash. In this case we now return a
valid resource scale because ClutterClone still expects cloned actors to
be realized, so in the worst case we'll just return the scale of the
stage.

Fixes: https://gitlab.gnome.org/GNOME/mutter/-/issues/808