clutter/text: Only queue relayout if the actor is already allocated
There's no reason to queue a relayout of the ClutterText when the
resource scale changed if it doesn't have a valid allocation, as it will
be allocated again anyway (taking the new resource scale into account,
too).

This fixes a few warnings when we try to queue a relayout during an
allocation cycle, as sometimes the resource scale is recomputed when we
request it inside clutter_text_allocate(), notifying the resource-scale
property and queing a relayout that is unnecessary (we're using the
correct resource scale for the current allocation already). This is
still not completely correct though, since the actor box passed to
current clutter_text_allocate() call still uses the dimensions that were
calculated using the old resource scale in
clutter_text_get_preferred_width/height().