clutter/actor: Invalidate stage-views on transformation changes
Since we now have the neccessary infrastructure to get notified about
changes to the absolute transformation matrix, we can also invalidate
the stage-views list on updates to this matrix.

So simply rename absolute_allocation_changed() to
absolute_position_or_size_changed() to make it clear this function is
not only about allocations, and call that function every time the
absolute position or size of an actor changes.

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1299