clutter/timeline: Use C99 types for durations and delays
Working with time gets a lot less confusing to work with when seeing the
actual size of the values your using, so use the C99 uint32_t type
instead of guint for durations and delays.

The reason we don't want uint64_t here (like usually when working with
time), is that we need to cast those values to double before dividing
them to get a fractional progress. Since casting uint64_t to double can
lead to precision issues, using uint32_t makes more sense here. This
still gives us a maximum duration of ~49 days, which should be enough
for most animations.