clutter/main: Update event actor after event was sitting in the queue
Apparently the source actor events can change while an event is sitting
inside the stage queue, for example when a window gets closed in
response to a wayland client handling an event.

Now when we're done queuing and compressing events and finally emit them
to Clutter, events can have an outdated source actor pointer, leading to
a crash.

To fix this, ask the stage for the current device actor again and update
the event source actor before emitting the event to Clutter.

Fixes: https://gitlab.gnome.org/GNOME/mutter/-/issues/2152