clutter: Emit events to ClutterActions before emitting them to actors
With the following commits, we'll introduce a mechanism for implicitly
grabbing the event handling ClutterActions of event sequences (ie. a
touchpoint or a button-press->release sequence).

For this to work, ClutterActions will be prioritized over traditional
ClutterActor event handling, and it will be possible to not emit the
event to actors at all if a ClutterAction has handled the event.

So as a first step, shuffle things around to emit events to
ClutterActions before emitting them to actors. We do this by
collecting the actions for an event to a reusable GPtrArray, and only
calling _clutter_actor_handle_event() if no ClutterAction handled the
event.

Given that we're going to make a clear distinction between sequence
events (ie. events that share context and together make up a single user
interaction) and discrete events, name the new functions accordingly
already.