clutter: Emit events to ClutterActions before emitting them to actors
As mentioned in a preceding commit message before this one, this
implements part 2) of what's needed for well working gesture handling in
Clutter, that is prioritizing ClutterActions over traditional
ClutterActor event handling.

In case an action has handled the event, it will not be forwarded to
actors and emitted on actors at all. Implement this in pretty much the
same way as we already did for collecting the event-handling actors,
using two GPtrArrays (one for sequence events that lives in the
PointerDeviceEntry and a generic one for all discrete events in
ClutterStagePrivate).

We're also going to make a clear distinction between emitting sequence
events (ie. events that share context and together make up a single user
interaction) and discrete events to actions, so name the new event
emission function accordingly.