clutter: Introduce implicit grabbing for delivering events to actions
Introduce everything that's needed to implicitly grab pointer/touch
events and deliver them to ClutterActions. The way it works is that we
collect a list of ClutterActions when the TOUCH_BEGIN/BUTTON_PRESS
happens, and from then on emit all events belonging to this touchpoint
or button press to those actions.

With this, ClutterActions are now guaranteed to see the whole sequence
of events for button presses or touchpoints. Since there are cases where
those points get taken away under our feet (stage grabs, TOUCH_CANCEL
events etc.), this also adds a way to notify actions that their implicit
grab has been lost.