clutter/actions: Implement new sequences_cancelled vfunc
Now that we have more robust API to get notified about points that got
cancelled, make use of it to cancel ClutterGestureActions and
ClutterClickActions.

Since the stage now is aware of our sequence grab, it will also cancel
that grab when a conflicting ClutterGrab appears, so we no longer have
to listen to GRAB_NOTIFY crossing events now.