clutter/gesture: Add a new ClutterAction for touch and mouse gestures
Introduce ClutterGesture, a new ClutterAction subclass and the successor
of ClutterGestureAction that brings the necessary tools to handle
sequences of events and abstract touch and mouse gestures from those.

The biggest difference compared to ClutterGestureAction is that
ClutterGesture provides the implementation with
points_added/moved/ended/cancelled events and expects the implementation
to move the ClutterGesture through the ClutterGestureState state
machine. This state machine is then used internally by ClutterGesture to
coordinate with other gestures.

With the next commits, ClutterGesture will handle relationships between
conflicting gestures completly by itself, allowing the implementation or
the user to specify the details of the relationship between two
gestures.

am first too