clutter/stage: Allow claiming sequence outside of event handling context
Another baby step just like the last commit: This commit takes care of
the opposite case: An action handling a sequence event stops further
emission of events to actors.

Since sequences remain around for longer than the context of just a
single event, it makes sense to provide a way to "claim" those sequences
even when outside of event handling context, so introduce API for that.

Note that crossing events are handled in a somewhat special way here:
Since there is no "right" or "wrong" with cases like this, it was
decided to hand all control over to actions themselves.