clutter/stage: Don't emit events to actors while gesture is recognizing
Since the last commits, it's now possible for gestures to "claim"
sequences by cancelling each other when one starts recognizing.

To get notified about gestures that started recognizing anywhere in the
actor tree, we can now use a little trick in ClutterStage: Create a
"fake" gesture called "recognizing_gestures_tracker" and attach this
gesture to all the sequence grabs that we create. This means the
recognizing_gestures_tracker will now track any gesture that's active in
the scenegraph and will get cancelled as soon as any gesture starts
recognizing.

We can make use of this to avoid emitting events to ClutterActors while
gestures are recognizing. It works by checking the trackers state: If
the state is POSSIBLE, it means that points were added to the tracker
and it's ready to recognize. We can interpret this as "no other gesture
has recognized yet", so in this case it's fine to send events to actors.
If the state isn't POSSIBLE, it means that the tracker was cancelled and
no longer has any points, in this case another gesture must have
cancelled it, so we avoid sending any events to the actor tree.

Note that this is handling new sequences getting added while a gesture
is running just fine: Since the last commit, ClutterGesture won't allow
the state change to POSSIBLE while any gesture is already RECOGNIZING.

am