clutter/stage: Don't emit crossing events while gesture is recognizing
Thanks to the stage now being aware of which sequences are claimed by
gestures, it's now also fairly straightforward to deal with crossing
events while a gesture is recognizing: As soon as something gets
claimed, we emit a LEAVE event to the scenegraph.

For mice, after a sequence grab ends, we also need to emit an ENTER
event again.