clutter/stage: Don't emit crossing events while gesture is recognizing
Thanks to the new recognizing_gestures_tracker on the stage, it's now
also fairly straightforward to deal with crossing events while a gesture
is recognizing: We'll just have to listen to the tracker moving to
CANCELLED state, that means another gesture must have recognized, so we
can now emit LEAVE events to the scenegraph.

For mice, after a sequence grab ends, we also need to emit an ENTER
event again.