clutter/gesture: Allow controlling relationships of independent gestures
Gestures are independent from each other when they are on different
actors and when they don't have a conflict over input with each other,
so for example a gesture on one window and a gesture on another window.

For those gestures (we call them "independent gestures") you don't want
to control how the conflicting input should be handled, i.e. whether one
gesture wins over another or whether both can be recognizing using the
same touchpoint (e.g. zoom+rotate on the same actor), instead you want
to control over whether they are allowed to start while another one is
running.

Cases where fine-grained control over this is needed is for example when
we want to allow two DND gestures to start at the same time, or when we
want to allow opening the overview while a DND gesture is recognizing.