clutter: Implicitly grab on button and touch event sequences
We'll soon introduce a new gesture tracking framework which heavily
depends on *all* gestures seeing all events.

For this to work, a larger change to event delivery is needed: Implicit
grabbing of all events for button and touch press->motion->release
sequences to ensure ClutterActions continue receiving events for the
whole sequence

This commit takes care of that: At the start of an event sequence we
collect all the event-handling actors to a GPtrArray that lives in
the PointerDeviceEntry, and then deliver all events belonging to that
sequence to the same actors until the sequence ends the time.

To avoid events getting pulled from under our feet when mutters event
filter returns CLUTTER_EVENT_STOP, this also introduces private API
(maybe_lost_sequence_grab()) on ClutterStage so that we can't end up
with stale sequences.

Note that this isn't handling ClutterGrabs correctly right now,
this will be a little tricky, so we'll take care of that in a future
commit.

To handle actors getting destroyed or unmapped during a grab, listen to
notify::grab on the deepmost actor in the sequence grab. This gives us a
notification when any actor inside the sequence grab goes unmapped.