clutter: Implicitly grab on button and touch event sequences
We'll soon introduce a new gesture tracking framework which heavily
depends on *all* gestures seeing all events.

For this to work, two larger changes to event delivery are needed:
1) Implicit grabbing of all events for button and touch
press->motion->release sequences to ensure ClutterActions continue
receiving events for the whole sequence
2) ClutterActions need to be prioritized over traditional ClutterActor
events and receive events before actors. This is to ensure ClutterActor
event handlers don't stop event delivery before all actions have seen
them.

This commit takes care of change 1): At the start of an event sequence
we collect all the event-handling actors to a GPtrArray that lives in
the PointerDeviceEntry, and then deliver all events belonging to that
sequence to the same actors until the sequence ends the time.

To avoid events getting pulled from under our feet when mutters event
filter returns CLUTTER_EVENT_STOP, this also introduces private API
(maybe_lost_sequence_grab()) on ClutterStage that we can't end up with
stale sequences.

Note that seat grabs are not handled in the "most correct" way right
now: When a seat grab happens while an implicit grab is in effect, the
implicit grab is not affected, even if the implicitly grabbed subtree is
not part of the seat-grabbed subtree. So the seat grab will only really
get enforced here once the press/touch sequence ends. While we could
also, on the start of seat grab, "strip out" the part of the implicit
grab tree that doesn't overlap with the seat grab tree, this would break
common use-cases, for example DND while also showing a popupMenu.

For the same reason actors getting unmapped will also continue receiving
events until the sequence ends, otherwise reparenting a drag actor when
DND starts would be impossible.