clutter/actor: Cache the visible paint volume
If no actors have changed their positions and we're only repainting
because eg. a window needs a repaint, the paint volumes of all actors
remain unchanged. There is no reason to redo those on every stage
update in case.

So introduce caching and invalidation logic for the visible_paint_volume
that allows us to avoid a ton of matrix multiplications that right now
are happening for the whole mapped actor tree on every redraw.

Note that this removes two places where the visible paint volume is set
to an empty paint volume: This is a compromise so that we can keep
around the cached pv when hiding and showing an actor, it does slightly
regress somewhere else though: When hiding -> moving -> showing an
actor, we'll now include the old pv of the actor in the redraw clip when
showing (even though that's not necessary), resulting in a bit of
overpaint. For the sake of simplicity let's not care about that, it
seems like a very.