clutter/actor: Always set stage view when there is just a single one
In case there is only one stage-view around, we can avoid doing all the
coordinate system conversions and intersection checking and simply set
the stage-views list to always use that single stage-view.

It does come with a drawback though: The behavior of the stage-views
list is now no longer perfectly consistent between the "there is only
one view" and "multiple views" cases: In the "there is only one view"
case we now no longer empty the list when the actor is off stage or
0-sized.

Considering that this is a really powerful optimization that has been
seen to shave off more than 1ms off the stage-update duration, this
seems like a compromise we should make.