clutter/stage: Cancel parts of sequence grabs when ClutterGrabs happen
A ClutterGrab takes precedence over sequence grabs, so when one happens,
let's check which part of the sequence grab is inside the new
ClutterGrab. Cancel and remove the parts which aren't, and if nothing
is in there anymore, cancel the whole sequence grab.

Emitting crossing events correctly here is getting quite tricky:

- When the sequence grab didn't get cancelled by the ClutterGrab, we
simply want to emit all GRAB_NOTIFY crossings to the sequence grab, as
we do with all other crossings.

- When the sequence grab did get cancelled and the new ClutterGrab wants
to emit ENTER crossings, we want those to be emitted to the actual
targets, so cancel the sequence grab before emission.

- In the last case where the sequence grab did get cancelled and the new
ClutterGrab wants to emit LEAVE crossings, those should be emitted to
the sequence grab again, so we cancel the grab only after the emission
of those.