window: Bail out of maximize functions when window can't maximize
So far we leave this check to the caller of meta_window_maximize(), if
the caller forgets to check for can_maximize(), the window will go into
a broken in-between state where the maximized-horizontally/vertically
properties are true, but the window isn't maximized.

Do the check inside maximize_internal() and bail out there, returning
FALSE before we set the properties.